% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data(file_path)
}
\arguments{
\item{file_path}{Path to data file}
}
\value{
A data frame
}
\description{
Import data
}
\examples{
\dontrun{
data <- import_data("path/to/file.csv")
}
}
