% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_gop}
\alias{plot_gop}
\title{Plot data of 'gop_2016' voting percentage and laws, mediated by historical party affiliation (color) and region (shapes)}
\usage{
plot_gop(data)
}
\arguments{
\item{data}{A scatter plot containing 'gop_2016' by 'laws', including 'historical' and 'region' data variables}
}
\value{
A scatter plot
}
\description{
Plot data of 'gop_2016' voting percentage and laws, mediated by historical party affiliation (color) and region (shapes)
}
