% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_state}
\alias{plot_state}
\title{Plot data of states and laws, mediated by historical party affiliation (color)}
\usage{
plot_state(data, top_n = 25)
}
\arguments{
\item{data}{A bar plot containing 'states' by 'laws', including the 'historical' data variable}
}
\value{
A bar plot
}
\description{
Plot data of states and laws, mediated by historical party affiliation (color)
}
