% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_urban}
\alias{plot_urban}
\title{Plot data of 'percent_urban' (percent of pop. that lives in an urban area) and laws, mediated by historical party affiliation (color) and region (shapes)}
\usage{
plot_urban(data)
}
\arguments{
\item{data}{A scatter plot containing 'percent_urban' by 'laws', including 'historical' and 'region' data variables}
}
\value{
A scatter plot
}
\description{
Plot data of 'percent_urban' (percent of pop. that lives in an urban area) and laws, mediated by historical party affiliation (color) and region (shapes)
}
